Change separator symbol from "_" to "-"
KM uses the dash "-" symbol as the standard char across the different
storage connectors, so this change aligns Postgres behavior with KM.
This is mostly visible when creating and fetching index names, because
"_" is automatically turned into "-", so the index name used in input
might differ from the one received in output, e.g. when listing
indexes.

Other:
- bump version to 0.3
- auto normalize "_" to "-" also in table name prefix config
- update docs and configs changing "km_" to "km-"
- clean up some test tables left over after running functional tests